Removed one sub directory that made project unable to run because imports were under a directory not expected
